[DFLASH] Drop redundant fp32 casts in the lattice einsum
The two .float() upcasts on the einsum inputs were defensive precision that
buys nothing: the correction is a small nudge added to fp32 unary_logits (the
add already promotes to fp32), sample_path re-casts transition_scores to fp32
before its softmax, and the verify is lossless rejection sampling (a slightly
different q only shifts accept length, never correctness). Removing them leaves
the einsum on the natural bf16 tensor-core path. Greedy accept is bit-identical
(t0 4.4912); a same-card A/B showed no throughput change either way -- this is a
readability cleanup, not a perf change.

+++ python/sglang/srt/models/dflash.py
@@ -683,8 +683,8 @@ class CandidateSelector(nn.Module):
         )
         corrections = torch.einsum(
             "blpr,blcr->blpc",
-            self.state_query(edge_inputs).float(),
-            candidate_factors[:, 1:].float(),
+            self.state_query(edge_inputs),
+            candidate_factors[:, 1:],
         ) / math.sqrt(self.state_rank)
         return candidate_ids, unary_logits, unary_logits[:, 1:].unsqueeze(2) + corrections
